[topgen] Emit a devicetree for each top
Add a devicetree emitter to topgen. The devicetree is a standard format
for describing a device to software, and it can be used in the future
with tools that provide abstractions and APIs to query the information
from the devicetree, enabling tests and testutils to support multiple
tops.

The amount of multi-top software support will depend on which APIs are
implemented, plus whether a test could fundamentally be shared.

Signed-off-by: Alexander Williams <awill@opentitan.org>